chore: move tests
These now belong in `models::uses`.
